+++ src/main/java/com/jinddung2/givemeticon/domain/item/service/ItemService.java
@@ -1,7 +1,7 @@
-package com.jinddung2.givemeticon.item.application;
+package com.jinddung2.givemeticon.domain.item.service;
 
-import com.jinddung2.givemeticon.item.domain.Item;
-import com.jinddung2.givemeticon.item.infrasturcture.mapper.ItemMapper;
+import com.jinddung2.givemeticon.domain.item.domain.Item;
+import com.jinddung2.givemeticon.domain.item.mapper.ItemMapper;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
