+++ src/main/resources/mapper/ItemMapper.xml
@@ -2,7 +2,7 @@
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="com.jinddung2.givemeticon.item.infrasturcture.mapper.ItemMapper">
+<mapper namespace="com.jinddung2.givemeticon.domain.item.mapper.ItemMapper">
     <insert id="save" parameterType="Item" useGeneratedKeys="true" keyProperty="id">
         INSERT INTO item (id,
                           brand_id,
