+++ src/test/java/com/jinddung2/givemeticon/item/facade/ItemCreationFacadeTest.java
@@ -3,9 +3,10 @@ package com.jinddung2.givemeticon.item.facade;
 import com.jinddung2.givemeticon.brand.application.BrandService;
 import com.jinddung2.givemeticon.brand.application.dto.BrandDto;
 import com.jinddung2.givemeticon.brand.domain.Brand;
-import com.jinddung2.givemeticon.item.application.ItemService;
-import com.jinddung2.givemeticon.item.domain.Item;
-import com.jinddung2.givemeticon.item.presentation.request.ItemCreateRequest;
+import com.jinddung2.givemeticon.domain.item.domain.Item;
+import com.jinddung2.givemeticon.domain.item.dto.request.ItemCreateRequest;
+import com.jinddung2.givemeticon.domain.item.facade.ItemCreationFacade;
+import com.jinddung2.givemeticon.domain.item.service.ItemService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
