+++ src/test/java/com/jinddung2/givemeticon/item/presentation/ItemControllerTest.java
@@ -3,8 +3,9 @@ package com.jinddung2.givemeticon.item.presentation;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.jinddung2.givemeticon.common.config.WebConfig;
 import com.jinddung2.givemeticon.common.interceptor.AuthInterceptor;
-import com.jinddung2.givemeticon.item.facade.ItemCreationFacade;
-import com.jinddung2.givemeticon.item.presentation.request.ItemCreateRequest;
+import com.jinddung2.givemeticon.domain.item.controller.ItemController;
+import com.jinddung2.givemeticon.domain.item.dto.request.ItemCreateRequest;
+import com.jinddung2.givemeticon.domain.item.facade.ItemCreationFacade;
 import com.jinddung2.givemeticon.user.application.LoginService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
